Promote flanking using math as a regular option
